Begin work on another representation of horizontal lines.
In this version, all corners or endpoints must be represented
by "*", except for the final (lower-right) endpoint, which
continues to be represented by "o"; and all vertical bars are
represented by "!". This means that a rectangle literal is
simply one huge prefix-expr.

This version (analog-punct.hpp) supports 1-D and 2-D literals,
but not yet 3-D literals.