Move the division by two later in the construction.
Types otoposidesobottom, osidesobottom and osidesobottomo
now use their "int s" template parameter to represent the
total number of vertical pipes in the construction, rather
than the number of vertical pipes divided by two. Something
like this will be necessary when adding cuboid constructions,
because we can have constructions like

    *--*
    !o--*
    *!  !
     *--o

which parses as otoposidesobottom(*--*!o--) * (*!!*--o)
with an "s" parameter of 1.

This doesn't change the observable behavior of the header.