Begin work on 3-D literals (cuboids).
This commit enables cuboids of dimensions WxHxD as long as
(H == D+1) && (D >= 1). Such constructions contain the substring
"*---**!!L", in which the top right corner of the front face is
immediately followed by the bottom left corner of the back face.
This is one family of cuboids that wasn't allowed by Eelis's
original constructions; but notice that I don't yet allow the
large family of cuboids that *was* allowed by Eelis.

This hasn't been exhaustively tested, either, so there are
certainly some combinations of o*-~|! that should be accepted
but aren't, even for cuboids of the proper dimensions.