Add a random test autogenerator for analog-punct.hpp.
This is a quick and dirty hack; it needs polishing.
But it does generate random valid rectangles and cuboids,
and the appropriate assertions to verify their properties.
A few changes to analog-punct.hpp later, and indeed all
the random tests pass!

analog-punct.hpp is still limited to cuboids with
(D >= 1) and (H == D+1), but I'm moving in the direction
of relaxing that to (H > D) with this commit.