Allow cuboids of dimension WxHxD for all (0 < D < H).
(D > 0) implies that the cuboid contains an "L".
(H > D) implies that the top of the front face is
encountered lexically before the rear corner of the
cuboid.

I've also expanded the autogenerator to generate these
cuboids, and to generate random 1-D lines, and to be
a little smarter about "random" characters that only
have one possibility in practice. For example, in the
sequence "| |" (vertical, whitespace, vertical), the
second vertical MUST be the "!" character, because
neither "! |" nor "| |" can be part of a valid C++
expression.